+++ lib/pipedrive/resources/participant.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 module Pipedrive
-  VERSION = "1.3.0"
+  class Participant < Resource
+  end
 end